Implement #6 - add game state
The gameState could be either "NONE", "SUCCEEDED" or "FAILED". "NONE"
is a default state. It is switched to "FAILED" from "NONE" when player
opened a cell with a mine or conducted chord at a cell with wrong flags
around.
"SUCCEEDED" state was not implemented this time.
The state is displayed beside the reset button.