Refactor openCells and minefield object
Re-removed the feature that precomputes connected components, in order
to implement flood fill same as minesweeper arbiter.
Flood fill is performed whenever an empty cell that have no adjacent
mine is opened.
Additionally, openCells traces how many cells have been opened so far,
and it is stored to minefield object.
It allows easy determining whether the game is cleared or not.